refactoring view recipe
