+++ code/recipe-book-app/src/components/index.js
@@ -12,7 +12,6 @@ import ImageGallery from "./ImageGallery";
 import Navigation from "./Navigation";
 import RecipeListCard from "./RecipeListCard";
 import SubPageHeader from "./SubPageHeader";
-import ViewRecipe from "./ViewRecipe";
 export {
 	RecipeForm,
 	FavoriteButton,
@@ -26,5 +25,4 @@ export {
 	Navigation,
 	RecipeListCard,
 	SubPageHeader,
-	ViewRecipe,
 };
